+++ chat.php
@@ -1,13 +1,10 @@
-<!DOCTYPE html>
-<html lang="zh-Hant">
-<head>
-    <meta charset="UTF-8">
-    <meta http-equiv="X-UA-Compatible" content="IE=edge">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/5.15.2/css/all.min.css" integrity="sha512-HK5fgLBL+xu6dm/Ii3z4xhlSUyZgTT9tuc/hSrtw6uzJOvgRr2a9jyxxT1ely+B+xFAmJKVSTbpM/CuL7qxO8w==" crossorigin="anonymous" />
-    <link rel="stylesheet" href="css/style.css">
-    <title>Chat App</title>
-</head>
+<?php 
+    session_start();
+    if (!isset($_SESSION['user_id'])) {
+        header("location: login.php");
+    }
+?>
+<?php include_once "header.php"; ?>
 <body>
     <div class="wrapper">
         <section class="chat-area">